// src/app/store/currentUser/currentUser.types.ts
export enum AuthActionTypes {
  SIGN_IN_START = '[AUTH] SIGN_IN_START',
  SIGN_IN_SUCCESS = '[AUTH] SIGN_IN_SUCCESS',
  SIGN_IN_FAILURE = '[AUTH] SIGN_IN_FAILURE',
  SIGN_UP_START = '[AUTH] SIGN_UP_START',
  SIGN_UP_SUCCESS = '[AUTH] SIGN_UP_SUCCESS',
  SIGN_UP_FAILURE = '[AUTH] SIGN_UP_FAILURE',
  LOGOUT = '[AUTH] LOGOUT',
}